refactor: PHPStan level 9 progress — 598→280 errors (318 fixed)
- Fixed traits (ShowsUpgradeBadge, LogsActivity) eliminating 96 multiplied errors
- Added typed array annotations for json_decode, config, webhook payloads
- Updated Collection generic types across services and pages
- Added null safety for tenant() calls in Filament pages
- Cast config()/settings() returns at usage points
- All 661 tests pass

+++ app/Actions/Tenants/CreateTenant.php
@@ -22,7 +22,7 @@ class CreateTenant
                 'name' => $user->name,
                 'email' => $user->email,
                 'plan' => SubscriptionTier::Starter->value,
-                'trial_ends_at' => now()->addDays(config('saas.trial_days', 30)),
+                'trial_ends_at' => now()->addDays((int) config('saas.trial_days', 30)),
                 'store_name' => $storeName,
                 'storefront_enabled' => $useKneadItStorefront,
                 'external_website' => $useKneadItStorefront ? null : $externalWebsite,
